Advanced Measure Layout -preliminary implementation
